Fix Battle Mage skills, fix AddDamParty MTS vs AddDamSkill2
- AddDamParty requires pOption (party ID) and wOption (char ID) to be set in order to work
    - I will use this for skills that explicitly mention the effect for party members in the description
- AddDamSkill increases damage for skills where skillID == rOption
- AddDamSkill2 just increases damage by nOption, use this for everything else

+++ src/main/java/net/swordie/ms/client/character/commands/AdminCommands.java
@@ -1168,7 +1168,7 @@ public class AdminCommands {
             for (short j : targetJobs) {
                 List<Skill> list = new ArrayList<>();
                 for (Skill skill : SkillData.getSkillsByJob(j)) {
-                    byte maxLevel = (byte) skill.getMaxLevel();
+                    int maxLevel = skill.getMaxLevel();
                     skill.setCurrentLevel(maxLevel);
                     skill.setMasterLevel(maxLevel);
                     list.add(skill);
